organized the components in the components folder and use underscores instead of dashes for the file names
