+++ frontend/src/components/navigation_bar/logout.js
