+++ frontend/src/components/navigation_bar/navigation_bar.css
