+++ frontend/src/components/note_actions/add_new_note.js
@@ -1,62 +1,58 @@
-/* Edit Note Page */
+/* Add New Note Page */
 
 import axios from "axios";
-import {useState, useEffect} from "react";
+import { useState } from "react";
+import { getToken } from "components/token/token";
 
-export const EditNoteForm = () => {
-	
-	let url = new URL(window.location.href);
-	const note_id = parseInt(url.searchParams.get("note_id"));
-
-	if (note_id === null) {
-		window.location.href = '/';
-	}
-
+export const AddNoteForm = () => {
 	const [title, setTitle] = useState('');
 	const [description, setDescription] = useState('');
 
 	const handleTitleChange = event => {
     	setTitle(event.target.value);
   	};
+
 	const handleDescriptionChange = event => {
     	setDescription(event.target.value);
   	};
+
 	const GoBack = () => {
 		window.location.href='/dashboard';
 	}
 
-	useEffect(() => {
-		/* Fetch Note */
-		axios.post(`http://localhost:8080/api/fetch-note`, {
-			note_id: note_id,
-		}).then((response) => {
-			if (note_id !== undefined) {
-				setTitle(response.data.title);
-				setDescription(response.data.description);
-			}
-		}).catch(e => {
-            console.log(e);
-        })
-	}, []);
-
 	const handleSubmit = async (e) => {
+		/* Add New Note Submission */
 		e.preventDefault();
-		axios.post(`http://localhost:8080/api/edit-note`, {
-			note_id: note_id,
-			title: title,
-			description: description,
-		}).then((response) => {
-			// redirect to the dashboard
-            window.location.href = '/dashboard';
-            console.log(response);
-		}).catch(e => {
-            console.log(e);
-        })
+		const token = getToken();
+		let user_id = undefined;
+		console.log(token)
+		if (token !== undefined) {
+			axios.post(`http://localhost:8080/api/get-decoded-token`, {
+				token: token,
+			}).then((response) => {
+				if (response.data.user_id !== undefined) {
+					user_id = response.data.user_id;
+					axios.post(`http://localhost:8080/api/add-new-note`, {
+						title: title,
+						description: description,
+						user_id: user_id,
+					}).then((response) => {
+						// redirect to the dashboard
+						window.location.href = '/dashboard';
+						console.log(response);
+					}).catch(e => {
+						console.log(e);
+					})
+				}
+			}).catch(e => {
+				console.log(e);
+			})
+		}
 	};
 
 	return (		
 		<div>
-			<h2> Edit Note </h2>
+			<h2> Add Note </h2>
 			<form method="post" onSubmit={handleSubmit}>
 				<div className="form-group">
 					<label htmlFor="exampleFormControlInput1">Title</label>
@@ -69,13 +65,14 @@ export const EditNoteForm = () => {
 				<div className="form-group">
 					<label htmlFor="exampleFormControlTextarea1">Description</label>
 					<textarea className="form-control" id="exampleFormControlTextarea1" 
-						name="description" rows="3" placeholder="Description"
+						name="description" rows="3" placeholder="Description" 
 						value={description}
 						onChange={handleDescriptionChange}
 						required> </textarea>
 				</div>
-
-				<button type="submit" className="btn btn-primary">Submit</button>
+				<button type="submit" className="btn btn-primary">
+					Submit
+				</button>
 				<button type="button" className="btn btn-secondary" 
 					onClick={ GoBack }>
 					Back
