+++ frontend/src/components/note_actions/delete_note.js
