+++ frontend/src/components/register/password-validator.js
