+++ frontend/src/components/view_notes/view_notes.css
